+++ sea-query-attr/tests/compile-fail/enum.rs
@@ -1,0 +1,8 @@
+use sea_query_attr::enum_def;
+
+#[enum_def]
+enum Hello {
+    Name,
+}
+
+fn main() {}