+++ src/extension/mysql/column.rs
@@ -2,7 +2,6 @@ use crate::Iden;
 
 #[derive(Debug, Copy, Clone)]
 pub enum MySqlType {
-    Blob,
     TinyBlob,
     MediumBlob,
     LongBlob,
@@ -11,7 +10,6 @@ pub enum MySqlType {
 impl Iden for MySqlType {
     fn unquoted(&self, s: &mut dyn std::fmt::Write) {
         let ty = match self {
-            Self::Blob => "blob",
             Self::TinyBlob => "tinyblob",
             Self::MediumBlob => "mediumblob",
             Self::LongBlob => "longblob",